requirements.txt
tensorflow
numpy
pillow 
django
djangorestframework 
gunicorn 